Move dashboard api from webapp to webservice to deprecate webapp in future
