+++ server/web-app/pom.xml
@@ -23,6 +23,12 @@
       <artifactId>spring-cloud-starter</artifactId>
     </dependency>
 
+    <!-- Feign, use feign to call remote service -->
+    <dependency>
+      <groupId>org.springframework.cloud</groupId>
+      <artifactId>spring-cloud-starter-openfeign</artifactId>
+    </dependency>
+
     <!-- Lombok -->
     <dependency>
       <groupId>org.projectlombok</groupId>
