+++ server/web-app/src/main/java/org/bithon/server/webapp/ui/LogController.java
@@ -16,14 +16,10 @@
 
 package org.bithon.server.webapp.ui;
 
-import jakarta.servlet.http.HttpServletResponse;
 import org.bithon.component.commons.Experimental;
-import org.bithon.server.storage.web.Dashboard;
 import org.bithon.server.webapp.WebAppModuleEnabler;
-import org.bithon.server.webapp.services.DashboardManager;
 import org.bithon.server.webapp.services.ServiceDiscovery;
 import org.springframework.context.annotation.Conditional;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -39,15 +35,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 public class LogController {
 
     private final ServiceDiscovery serviceDiscovery;
-    private final DashboardManager dashboardManager;
 
-    public LogController(ServiceDiscovery serviceDiscovery, DashboardManager dashboardManager) {
+    public LogController(ServiceDiscovery serviceDiscovery) {
         this.serviceDiscovery = serviceDiscovery;
-        this.dashboardManager = dashboardManager;
     }
 
     @GetMapping("/web/log")
-    public String log(Model model, HttpServletResponse servletResponse) {
+    public String log(Model model) {
         model.addAttribute("apiHost", serviceDiscovery.getApiHost());
         model.addAttribute("store", "");
         model.addAttribute("storeDisplayName", "");
@@ -55,18 +49,11 @@ public class LogController {
     }
 
     @GetMapping("/web/log/{store}")
-    public String log(@PathVariable("store") String store,
-                      Model model,
-                      HttpServletResponse servletResponse) {
-        Dashboard dashboard = dashboardManager.getDashboard(store);
-        if (dashboard == null) {
-            servletResponse.setStatus(HttpStatus.NOT_FOUND.value());
-            return null;
-        }
+    public String log(@PathVariable("store") String store, Model model) {
 
         model.addAttribute("apiHost", serviceDiscovery.getApiHost());
         model.addAttribute("store", store);
-        model.addAttribute("storeDisplayName", dashboard.getMetadata() == null ? store : dashboard.getMetadata().getTitle());
+        model.addAttribute("storeDisplayName", store);
         return "log/detail";
     }
 }
